+++ internal/provider/resource/node/meta.go
@@ -11,9 +11,11 @@ import (
 )
 
 // Ensure provider defined types fully satisfy framework interfaces
-var _ resource.Resource = &NodeResource{}
-var _ resource.ResourceWithImportState = &NodeResource{}
-var _ resource.ResourceWithModifyPlan = &NodeResource{}
+var (
+	_ resource.Resource                = &NodeResource{}
+	_ resource.ResourceWithImportState = &NodeResource{}
+	_ resource.ResourceWithModifyPlan  = &NodeResource{}
+)
 
 type NodeResource struct {
 	cfg *common.ProviderConfig
@@ -30,7 +32,14 @@ func (r *NodeResource) Configure(ctx context.Context, req resource.ConfigureRequ
 func (r *NodeResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	// This description is used by the documentation generator and the language
 	// server.
-	resp.Schema.MarkdownDescription = "A node resource represents a CML node. At create time, the lab ID, a node definition and a label must be provided.  Other attributes are optional.  Note that some attributes can't be changed after the node state has changed to `STARTED` (see the lifecyle resource) once. Changing attributes will then require a replace."
+	resp.Schema.MarkdownDescription = "A node resource represents a CML node. At create time, the lab ID, a " +
+		"node definition and a label must be provided.  Other attributes are optional.  Note that some " +
+		"attributes can't be changed after the node state has changed to `STARTED` (see the `lifecyle` resource) " +
+		"once. Changing attributes will then require a replace.  " +
+		"Node configurations are \"day zero\" configurations. Replacing a configuration typically requires a " +
+		"node replacement if the node has been started.  No Configurations can be provided for unmanaged switches. " +
+		"External connectors require the connector label (like \"NAT\"), not the device name (like \"virbr0\"). " +
+		"The available connectors can be retrieved via the external connector data source."
 	resp.Schema.Attributes = cmlschema.Node()
 	resp.Diagnostics = nil
 }